data_engine/requirements.txt
google-cloud-bigquery
google-cloud-storage
google-cloud-aiplatform
praw
pandas
pyarrow
python-dotenv
functions-framework
requests
openpyxl
pytrends
